Fixed compiling the library with AS_NO_CLASS_METHODS on gnuc
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@2866 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/docs/doxygen/source/angelscript.h
@@ -760,7 +760,7 @@ struct asSFuncPtr
 {
 	asSFuncPtr(asBYTE f)
 	{
-		for( int n = 0; n < sizeof(ptr.dummy); n++ )
+		for( size_t n = 0; n < sizeof(ptr.dummy); n++ )
 			ptr.dummy[n] = 0;
 		flag = f;
 	}
